fix: rename replace home components
