+++ components/Home/HomeHead.tsx
@@ -1,11 +1,11 @@
-import Head from 'next/head'
+import Head from "next/head";
 
-export default function PageHead() {
+export default function HomeHead() {
   return (
     <Head>
       <title>webgam</title>
       <meta name="description" content="webgam's home page" />
       <link rel="icon" href="/favicon.ico" />
     </Head>
-  )
+  );
 }
