+++ components/Home/HomePage.module.scss
@@ -1,223 +1,2 @@
-@use "../../styles/colors";
-
-.wrapper {
-  width: 100vw;
-  background: url("/images/home_background.png") repeat;
-  padding: 72px 0;
-  position: relative;
-  z-index: 0;
-
-  .container {
-    width: 1130px;
-    margin: auto;
-    text-align: center;
-
-    .homeHeader {
-      display: flex;
-      justify-content: space-between;
-
-      .moreButton {
-        margin-top: 10px;
-      }
-
-      .user {
-        display: flex;
-        align-items: center;
-        gap: 20px;
-
-        .signButtons {
-          display: flex;
-          gap: 20px;
-
-          button {
-            color: colors.$primary;
-          }
-        }
-
-        .profile {
-          width: 45px;
-          height: 45px;
-          border-radius: 50%;
-          background: colors.$mediumgray;
-          box-shadow: 4px 4px 4px rgba(0, 0, 0, 0.1);
-        }
-      }
-    }
-
-    .homeSidebar{
-      position: fixed;
-      left:0;
-      top:0;
-      width:300px;
-      height:100%;
-      display: flex;
-      flex-direction: column;
-      background-color: white;
-      z-index: 999;
-
-      .closeButton{
-        position: absolute;
-        top:0;
-        right:15px;
-      }
-
-      .label{
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        height:60px;
-        border-top:1px solid #cecece;
-        border-bottom:1px solid #cecece;
-      }
-      .projectList{
-        .projectItem{
-          display: flex;
-          align-items: center;
-          padding-left:30px;
-          color:gray;
-          height:60px;
-          border-bottom:1px solid #cecece;
-         cursor: pointer;
-          &:hover{
-           background-color: bisque;
-          }
-        }
-      }
-
-      .addButton{
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        min-height:60px;
-        padding:20px;
-        border-bottom:1px solid #cecece;
-        &:hover{
-          background-color: lightgrey;
-        }
-      }
-      .addForm{
-        display: flex;
-        align-items: center;
-        justify-content: space-around;
-        button{
-          border:1px solid black;
-        }
-      }
-    }
 
 
-    .logo {
-      margin-top: 69px;
-    }
-
-    .bannerContainer {
-      margin: auto;
-      margin-top: 147px;
-      width: 780px;
-      height: 233px;
-      box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
-      border-radius: 202px 66px 179px 61px;
-      overflow: hidden;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      position: relative;
-
-      .banner {
-        width: 100%;
-        height: auto;
-        position: absolute;
-        z-index: -1;
-      }
-
-      button {
-        position: absolute;
-        top: 77px;
-
-        &.left {
-          left: 0;
-          transform: rotate(180deg);
-        }
-
-        &.right {
-          right: 0;
-        }
-      }
-    }
-
-    .cardList {
-      margin-top: 220px;
-      width: 1120px;
-      display: grid;
-      column-gap: 28px;
-      row-gap: 96px;
-      grid-template-columns: repeat(4, 1fr);
-
-      .cardItem {
-        cursor: pointer;
-        transition: transform 0.3s ease-out;
-        position: relative;
-
-        .cardMark {
-          position: absolute;
-          z-index: 1;
-          left: 9px;
-          top: -10px;
-          text-align: center;
-
-          .text {
-            position: absolute;
-            top: 5px;
-            left: 5px;
-
-            span {
-              left: 2px;
-              color: white;
-              display: block;
-            }
-          }
-        }
-
-        &:hover {
-          transform: translate(0, -15px);
-        }
-
-        .cardImage {
-          height: 139px;
-          width: 261px;
-          box-shadow: 5px 5px 7px 0px #00000033;
-          filter: drop-shadow(5px 5px 7px rgba(0, 0, 0, 0.2));
-          border-radius: 0px 35px 0px 0px;
-          overflow: hidden;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-
-          img {
-            width: 100%;
-            height: auto;
-          }
-        }
-
-        .cardInfo {
-          height: 120px;
-          width: 261px;
-          left: 395px;
-          box-shadow: 5px 5px 7px rgba(0, 0, 0, 0.2);
-          border-radius: 0px 0px 0px 35px;
-          color: white;
-          padding: 16px 32px;
-
-          h2 {
-            margin: 0;
-          }
-
-          div {
-            margin-top: 4px;
-          }
-        }
-      }
-    }
-  }
-}
