+++ components/Home/Logo/Logo.module.scss
@@ -1,0 +1,4 @@
+
+.logo {
+  margin-top: 69px;
+}
