--- pages/drag-and-drop/DragAndDrop.module.scss
@@ -1,6 +1,0 @@
-.container {
-  position: absolute;
-  width:100%;
-  height:100%;
-}
-
