--- pages/drag-and-drop/index.tsx
@@ -1,15 +1,0 @@
-import DynamicCanvas from "../../components/dev/DynamicCanvas/DynamicCanvas";
-
-import StaticContainer from "../../components/dev/StaticContainer/StaticContainer";
-import Modals from "../../components/dev/Modal/Modals";
-
-const DragAndDrop = () => {
-  return (
-    <>
-      <StaticContainer />
-      <DynamicCanvas />
-    </>
-  );
-};
-
-export default DragAndDrop;
